pyproject.toml
[tool.poetry]
name = "faucet-task"
version = "0.1.0"
description = ""
authors = ["Vasiliy <lmhex@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
web3 = "^7.3.0"
Django = "^5.1.1"
djangorestframework = "^3.15.2"
django-redis = "^5.4.0"
django-environ = "^0.11.2"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
